1.0.0 update
Vortex will now give the workshop a consistent ID, so if you import the same workshop mod again it'll overwrite the original.
Added a new column to the import table to show if you've already imported the mod before.
Added some help text to the bottom of the cleanup stage to allow you to manually open your Steam Workshop directory if you get stuck.